requirements.txt
google-genai>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.0.0
pypdf>=5.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
streamlit>=1.44.0
